Refactor all networking into networking package
This refactors all networking client behavior into the networking
package. The network client has been removed from the compute Service
struct, instead, we use the networking Service. Ports and trunking logic
have been moved out of instance.go in the compute package into port.go
and trunk.go in the networking package.

The process for refactoring is simple enough:

1. Delete network client from compute package
2. Move any functions using the network client to the networking package
3. Refactor the functions to call the networking client interface
instead of gophercloud directly
