+++ pkg/cloud/services/networking/securitygroups.go
@@ -332,6 +332,34 @@ func (s *Service) generateDesiredSecGroups(openStackCluster *infrav1.OpenStackCl
 	return desiredSecGroups, nil
 }
 
+func (s *Service) GetSecurityGroups(securityGroupParams []infrav1.SecurityGroupParam) ([]string, error) {
+	var sgIDs []string
+	for _, sg := range securityGroupParams {
+		listOpts := groups.ListOpts(sg.Filter)
+		if listOpts.ProjectID == "" {
+			listOpts.ProjectID = s.projectID
+		}
+		listOpts.Name = sg.Name
+		listOpts.ID = sg.UUID
+		SGList, err := s.client.ListSecGroup(listOpts)
+		if err != nil {
+			return nil, err
+		}
+
+		if len(SGList) == 0 {
+			return nil, fmt.Errorf("security group %s not found", sg.Name)
+		}
+
+		for _, group := range SGList {
+			if isDuplicate(sgIDs, group.ID) {
+				continue
+			}
+			sgIDs = append(sgIDs, group.ID)
+		}
+	}
+	return sgIDs, nil
+}
+
 func (s *Service) DeleteSecurityGroups(openStackCluster *infrav1.OpenStackCluster, clusterName string) error {
 	secGroupNames := []string{
 		getSecControlPlaneGroupName(clusterName),
@@ -553,4 +581,16 @@ func convertOSSecGroupRuleToConfigSecGroupRule(osSecGroupRule rules.SecGroupRule
 		RemoteGroupID:   osSecGroupRule.RemoteGroupID,
 		RemoteIPPrefix:  osSecGroupRule.RemoteIPPrefix,
 	}
+}
+
+func isDuplicate(list []string, name string) bool {
+	if len(list) == 0 {
+		return false
+	}
+	for _, element := range list {
+		if element == name {
+			return true
+		}
+	}
+	return false
 }
